web/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
qrcode==7.4.2
python-dotenv==1.0.0
Pillow==10.2.0
